+++ packages/automation-contracts/scheduler/hardhat.config.js
@@ -12,7 +12,7 @@ require("hardhat/config");
  */
 module.exports = {
     solidity: {
-        version: "0.8.19",
+        version: "0.8.23",
         settings: {
             optimizer: {
                 enabled: true,