+++ packages/ethereum-contracts/hardhat.config.ts
@@ -97,7 +97,7 @@ function createNetworkConfig(
 
 const config: HardhatUserConfig = {
     solidity: {
-        version: "0.8.19",
+        version: "0.8.23",
         settings: {
             optimizer: {
                 enabled: true,
@@ -182,7 +182,7 @@ const config: HardhatUserConfig = {
     },
     mocha: {
         timeout: 250000,
-        parallel: !!process.env.HARDHAT_RUN_PARALLEL,
+        parallel: !!+process.env.HARDHAT_RUN_PARALLEL,
         jobs: process.env.HARDHAT_TEST_JOBS
             ? parseInt(process.env.HARDHAT_TEST_JOBS)
             : undefined,