+++ packages/ethereum-contracts/test/foundry/superfluid/ERC721.t.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: AGPLv3
-pragma solidity 0.8.19;
+pragma solidity 0.8.23;
 
 import { IERC721Metadata } from "@openzeppelin/contracts/interfaces/IERC721Metadata.sol";
 import { FoundrySuperfluidTester } from "../FoundrySuperfluidTester.sol";