+++ packages/ethereum-contracts/CHANGELOG.md
@@ -41,6 +41,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixes
 - [`dev-scripts/deploy-test-framework.js`](dev-scripts/deploy-test-framework.js) compatible with both ethers-v5 and ethers-v6 now
+- `distribute` and `distributeWithCtx` API made consistent
 
 ## [v1.8.1] - 2023-08-28