+++ packages/ethereum-contracts/CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+### Added
+- bump solc to 0.8.23
+
 ## [v1.9.0] - 2024-01-09
 
 ### Breaking